refactor: 중복 ApiMappingItem 제거 및및 공통 DTO import 사용
