+++ ai/dto/response/apilist/apilist_validation_response.py
@@ -1,18 +1,9 @@
-from typing import List, Dict, Any
+from typing import List
 from pydantic import BaseModel
 from datetime import datetime
-
-class ApiMappingItem(BaseModel):
-    scenarioId: str
-    stepName: str
-    apiName: str
-    description: str
-    url: str
-    method: str
-    parameters: Dict[str, Any]
-    responseStructure: Dict[str, Any]
+from dto.request.apilist.common import ApiMappingItem
 
 class ApiListValidationResponse(BaseModel):
-    processedAt: str
+    processedAt: datetime
     validationRate: float
     apiMapping: List[ApiMappingItem]
